feat(driver): add ioctl support

+++ driver/modern_bpf/programs/tail_called/events/syscall_dispatched_events/ioctl.bpf.c
@@ -25,15 +25,15 @@ int BPF_PROG(ioctl_x, struct pt_regs *regs, long ret) {
 	ringbuf__store_s64(&ringbuf, ret);
 
 	/* Parameter 2: fd (type: PT_FD) */
-	int32_t fd = (int32_t)extract__syscall_argument(regs, 0);
-	ringbuf__store_s64(&ringbuf, (int64_t)fd);
+	int64_t fd = (int64_t)(int32_t)extract__syscall_argument(regs, 0);
+	ringbuf__store_s64(&ringbuf, fd);
 
 	/* Parameter 3: request (type: PT_UINT64) */
-	uint64_t request = extract__syscall_argument(regs, 1);
+	uint64_t request = (uint64_t)extract__syscall_argument(regs, 1);
 	ringbuf__store_u64(&ringbuf, request);
 
 	/* Parameter 4: argument (type: PT_UINT64) */
-	uint64_t argument = extract__syscall_argument(regs, 2);
+	uint64_t argument = (uint64_t)extract__syscall_argument(regs, 2);
 	ringbuf__store_u64(&ringbuf, argument);
 
 	/*=============================== COLLECT PARAMETERS  ===========================*/
